\name{cbind.arrayData}
\alias{cbind.arrayData}
\title{Cbind for objects of class arrayData}
\description{Cbind for objects of class arrayData,
  cf. \code{\link{arrayData-class}}, you may simply
  use cbind instead of cbind.arrayData .}
\usage{cbind.arrayData(...)}
\arguments{
  \item{...}{\code{\link{arrayData}} objects}
}

\details{  cf. \code{\link{arrayData-class}}  }

\value{an object of class \code{\link{arrayData}}}

\author{Andreas Buness <a.buness@dkfz.de>}


\seealso{ \code{\link{cbind}}, 
          \code{\link{arrayData-class}}
        }
\examples{}
\keyword{methods}
