\name{cbind.exprSetRG}
\alias{cbind.exprSetRG}
\title{Cbind for objects of class exprSetRG}
\description{Cbind for objects of class exprSetRG,
  cf. \code{\link{exprSetRG-class}}, you may simply
  use cbind instead of cbind.exprSetRG.}
\usage{cbind.exprSetRG(...)}
\arguments{
  \item{...}{ \code{\link{exprSetRG}} objects}
}

\details{  cf. \code{\link{exprSetRG-class}}  }

\value{an object of class \code{\link{exprSetRG}} }

\author{Andreas Buness <a.buness@dkfz.de>}


\seealso{ \code{\link{cbind}}, 
          \code{\link{exprSetRG-class}}
        }
\examples{}
\keyword{methods}
