%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% This file was automatically generated by Rdoc on 2006-04-12  8:38:17.
% Do NOT modify this file, instead modify the source (/buness/R/TestingOfarrayMagic/Rpacks/arrayMagic/R/interweave.R).
% Rdoc is copyright (C) 2001, Henrik Bengtsson, henrikb@braju.com.
% Download the Rdoc compiler at http://www.braju.com/R/
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%


 \name{interweave}

 \title{interweave}

 \alias{interweave}

 \description{
          The first object is "interweaved" with the second object
          and the resulting object with the following object and
          continued as long as the last object is reached.
          Note: \code{AsIs} class objects are coerced to a vector
                before interweaving starts.}

 \value{An "interweaved" object, cf. example.}

 \usage{interweave(...)}

 \arguments{
  \item{...}{vectors, matrices, one- and two-dimensional arrays;
             an \code{AsIs} class object is coerced to a vector.}
 }
 
 \examples{
    x <- seq(1,100,2)
    y <- seq(2,100,2)
    z <- interweave(x,y)
    stopifnot( all.equal(z,1:100) )
	}

 \keyword{manip}

 \author{Andreas Buness <a.buness@dkfz.de>}
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% End of File
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
