%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% This file was automatically generated by Rdoc on 2006-04-12  8:38:17.
% Do NOT modify this file, instead modify the source (/buness/R/TestingOfarrayMagic/Rpacks/arrayMagic/R/plot.imageMatrix.R).
% Rdoc is copyright (C) 2001, Henrik Bengtsson, henrikb@braju.com.
% Download the Rdoc compiler at http://www.braju.com/R/
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%


 \name{plot.imageMatrix}

 \title{Visualisation of a matrix}

 \alias{plot.imageMatrix}

 \keyword{hplot}

 \description{Visualisation of a data matrix, e.g. a matrix representing
  a two dimensional area, a matrix of distance or similarity scores,
  or any numeric matrix. Coloured boxes represent values, labels will
  be drawn inside the box.}

 \usage{plot.imageMatrix(x, labelMatrix, zlim, separateZScale=FALSE, zScale=TRUE, colourRamp, reverseYaxis = TRUE, labels, xLabels, yLabels, width=8, height=7, labelMatrixTextScaling=0.7, plotOutput="standard", fileName, ...)}

 \arguments{
  \item{x}{data matrix, class \code{matrix}; required; default missing}
  \item{labelMatrix}{matrix of labels of class \code{array} of dim(\code{x});
                     optional; default missing}
  \item{zlim}{numeric vector; defines the maximal z-range of the plot;
              default missing}
  \item{zScale}{logical; adds a scale to the plot; required; default TRUE}
  \item{separateZScale}{logical; extra plot of the scale; required; default FALSE}
  \item{colourRamp}{vector of colours; optional; default missing}
  \item{reverseYaxis}{logical; required; default \code{TRUE}}
  \item{labels}{vector of names corresponding to  a quadratic \code{x};
                optional; default missing}
  \item{xLabels}{vector of names corresponding to the columns of
                 \code{x} overrides \code{labels}; optional; default missing}
  \item{yLabels}{vector of names corresponding to the rows of
                 \code{x} overrides \code{labels}; optional; default missing}
  \item{width}{graphics window width; required; default: 8}
  \item{height}{graphics window height; required; default: 7}
  \item{labelMatrixTextScaling}{numeric; required; default: 0.7}

  \item{plotOutput}{character string specifying
          either "standard", "screen", "twoScreens", "pdf" or "win.metafile";
          required; default: "standard"}
  \item{fileName}{character string specifying the file path and file name;
                  optional; default missing}
  \item{...}{arguments are passed to \code{image}(), for example
             \code{xlab}, \code{ylab} (x- and y-axis label) and
             \code{main} (plot title) and \code{asp=1} to set the
             aspect ratio to 1}
 }

 \details{If no labels are supplied, the dimnames of \code{x} or if missing
          a simple numbering is used instead.}

 \examples{


 plot.imageMatrix(x=matrix(c(3,4,4,3),nrow=2, ncol=2),labels=c("one","two"))
 ma <- matrix(c(0.3,0.01,0.7,0.1,0.5,0.3,1,0.5,01), nrow=3,ncol=3)
 class(ma) <- c("imageMatrix", "matrix")
 plot(ma, labelMatrix=ma,labels=c("one","two", "three"), zlim=c(0,1))
 ma <- matrix(c(0.3, 0.01, 0.7, 0.1, 0.5, 0.3, 1, 0.5, 1, 0, 1, 0), nrow = 4, ncol = 3, byrow=TRUE )
 class(ma) <- c("imageMatrix", "matrix")
 plot(ma, labelMatrix = ma, xLabels = c("one", "two", "three"), zlim=c(0,1))
 plot(ma, reverseYaxis=FALSE, labelMatrix = ma, xLabels = c("one", "two", "three"), zlim=c(0,1))

\dontshow{

 plot.imageMatrix(main="Similarities",x=matrix(c(3,4,4,3),nrow=2, ncol=2),labels=c("one","two"))
 ma <- matrix(c(0.3,0.01,0.7,0.1,0.5,0.3,1,0.5,01), nrow=3,ncol=3)
 plot.imageMatrix(main="Similarities",x=ma, labelMatrix=ma,labels=c("one","two", "three"), zlim=c(0,1))
 ma <- matrix(c(0.3, 0.01, 0.7, 0.1, 0.5, 0.3, 1, 0.5, 1, 0, 1, 0), nrow = 4, ncol = 3, byrow=TRUE )
 plot.imageMatrix(main = "x", x = ma, labelMatrix = ma, xLabels = c("one", "two", "three"), zlim=c(0,1) )
 plot.imageMatrix(matrix(1:10,nrow=1))
}


 }

 \author{Andreas Buness <a.buness@dkfz.de>}

 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% End of File
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
