%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% This file was automatically generated by Rdoc on 2006-04-12  8:38:17.
% Do NOT modify this file, instead modify the source (/buness/R/TestingOfarrayMagic/Rpacks/arrayMagic/R/processArrayDataObject.R).
% Rdoc is copyright (C) 2001, Henrik Bengtsson, henrikb@braju.com.
% Download the Rdoc compiler at http://www.braju.com/R/
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%


 \name{processArrayDataObject}


 \title{Automated processing and normalisation of an arrayData-object
        (part of function processArrayData)}

 \alias{processArrayDataObject}

 \keyword{utilities}

 \description{ Automated processing and normalisation of an
               \code{\link{arrayData}}-object.
               The function is called
               by \code{\link{processArrayData}}.
	}

 \value{
        A list of objects, i.e.  an "exprSetRGObject" and 
        "arrayDataObject"
        with corresponding class types
        \code{\link{exprSetRG-class}},
        \code{\link{arrayData-class}}.
       }

 \usage{processArrayDataObject(arrayDataObject, spotIdentifier = "Name", verbose = TRUE, normalisationMethod = "vsn", subtractBackground=FALSE, spotsRemovedBeforeNormalisation = NULL, spotsRemovedAfterNormalisation = NULL, subGroups = NULL, channelsSeparately = FALSE, hybridisationGroups = NULL)}

 \arguments{

  \item{arrayDataObject}{object of class \code{\link{arrayData-class}};
                         required; default missing
                        }



  \item{spotIdentifier}{ character string; required; default "Name".
                         \code{spotIdentifier} specifies the column
                         in the image analysis result files which contain
                         spot or gene identifiers.}
  \item{verbose}{ logical; required; default: \code{TRUE} }



  \item{normalisationMethod}{character string; required; default: "vsn";
        cf. argument \code{method} of \code{\link{normalise}} }
  \item{subtractBackground}{ logical; required; default: \code{FALSE}
        cf. \code{\link{normalise}} }
  \item{spotsRemovedBeforeNormalisation}{ vector of character strings;
       required; default \code{NULL}.
       All spots which match the string(s) will be excluded  
       already before normalisation and will not be present
       in the resulting objects at all.
       }
  \item{spotsRemovedAfterNormalisation}{ vector of character strings;
       required; default: \code{NULL};
       cf. argument \code{spotsRemovedBeforeNormalisation}
       }
  \item{subGroups}{character string or \code{NULL};
                   required; cf. \code{\link{normalise}};
                   default: \code{NULL} }
  \item{channelsSeparately}{logical; required;
                            cf. \code{\link{normalise}};
                            default: \code{FALSE}}
  \item{hybridisationGroups}{list of numeric vectors;
                            cf. \code{\link{normalise}};
                            required; default: \code{NULL}}

 }

 \examples{
       LOADPATH <- file.path(.path.package("arrayMagic"), "extdata")
 	SLIDEDESCRIPTIONFILE <- "slideDescription"

       slideDescription <- readpDataSlides(
                             loadPath=LOADPATH, 
                             slideDescriptionFile=SLIDEDESCRIPTIONFILE
                           )
 	arrayDataObject <- readIntensities(
                             loadPath=LOADPATH,
                             slideDescription=slideDescription
                          )

       resultList <- processArrayDataObject( arrayDataObject=arrayDataObject )
         
 }
 

 \seealso{ \code{\link{processArrayData}},
           \code{\link{normalise}},
           \code{\link{exprSetRG-class}},
           \code{\link{arrayData-class}}
         }

 \author{Andreas Buness <a.buness@dkfz.de>}

 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% End of File
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
