%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% This file was automatically generated by Rdoc on 2006-04-12  8:38:17.
% Do NOT modify this file, instead modify the source (/buness/R/TestingOfarrayMagic/Rpacks/arrayMagic/R/qualityDiagnostics.R).
% Rdoc is copyright (C) 2001, Henrik Bengtsson, henrikb@braju.com.
% Download the Rdoc compiler at http://www.braju.com/R/
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%


 \name{qualityDiagnostics}

 \title{Generates quality diagnostic plots}

 \alias{qualityDiagnostics}

 \usage{qualityDiagnostics(arrayDataObject, exprSetRGObject, qualityParametersList, groupingColumn, slideNameColumn, savePath=".", completeOutput=FALSE, verbose=TRUE, plotOutput="pdf")}

 \description{Several quality diagnostic plots are generated.
              The distributions of the (normalised) intensities,
              as well as overall similarities between hybridisations and
              several other quality parameters are graphically visualised
              (cf. function \code{\link{qualityParameters}}).
             }

 \value{The function is called for its side effect.}


 \arguments{
  \item{arrayDataObject}{\code{\link{arrayData-class}};
                         required; default: missing}
  \item{exprSetRGObject}{\code{\link{exprSetRG-class}};
                         required, default: missing}
  \item{qualityParametersList}{the return value of
                               \code{\link{qualityParameters}}
                               required; default: missing}

  \item{groupingColumn}{character string; optional; default: missing;
                         specifies a column of
                         \code{pDataSlide(exprSetRGObject)} which is
                         used to indicate the hybridisation groups
                         in the clustered similarity plots.}
  \item{slideNameColumn}{character string; optional; default: missing;
                         specifies a column of
                         \code{pDataSlide(exprSetRGObject)}.}
  \item{savePath}{character string; required; default: "."}
  \item{completeOutput}{required; default: \code{FALSE}}
  \item{verbose}{ logical; required; default: \code{TRUE}}
  \item{plotOutput}{character string specifying the output type;
        either "screen", "pdf" or "win.metafile"; default: "pdf"}
 }
 
 \details{}

 \seealso{\code{\link{qualityParameters}},
          \code{\link{exprSetRG-class}},
          \code{\link{arrayData-class}}
         }

 \examples{
	}

 \keyword{hplot}

 \author{Andreas Buness <a.buness@dkfz.de>}
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% End of File
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
