%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% This file was automatically generated by Rdoc on 2006-04-12  8:38:18.
% Do NOT modify this file, instead modify the source (/buness/R/TestingOfarrayMagic/Rpacks/arrayMagic/R/readpDataSlides.R).
% Rdoc is copyright (C) 2001, Henrik Bengtsson, henrikb@braju.com.
% Download the Rdoc compiler at http://www.braju.com/R/
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%


 \name{readpDataSlides}

 \title{Reads the description of all slides or channels}

 \alias{readpDataSlides}

 \keyword{IO}

 \description{The function reads the \code{slideDescriptionFile},
              which contains all information row-wise
              for each slide/microarray
              or for each channel in a tab-deliminated text file.
              The file is read up to the first line
              containing solely "white space".
	}

 \value{ \code{\link{data.frame}} }

 \usage{readpDataSlides(slideDescriptionFile = "slideDescription.txt", loadPath = ".", deleteBlanks = TRUE, verbose = TRUE)}

 \arguments{

  \item{slideDescriptionFile}{ character string;
       required; default "slideDescription.txt".
       The first line of the file must contain all column names,
       in particular the column named "fileName" which contains
       image analyis result file names and possibly
       additionally a column which contains the slide names;
       cf. the arguments \code{fileNameColumn} and
                         \code{slideNameColumn}
           of \code{\link{readIntensities}}.
       Note: Column names must not contain underlines or spaces.
       }
  \item{loadPath}{ character string; required; default: ".".
       The path is used to load the \code{slideDescriptionFile};
       note: "." refers to the working directory.
       }
  \item{deleteBlanks}{ logical; required; default: \code{TRUE}.
                       Any blank character is removed from
                       the text body of the
                       \code{slideDescriptionFile}.
                       Thus no item/name should contain spaces.
                     }
  \item{verbose}{ logical; required; default: \code{TRUE} }

 }

 \examples{
       LOADPATH <- file.path(.path.package("arrayMagic"), "extdata")
 	SLIDEDESCRIPTIONFILE <- "slideDescription"
  
 	resultObject <- readpDataSlides(
                             loadPath=LOADPATH, 
                             slideDescriptionFile=SLIDEDESCRIPTIONFILE
                  )
 }

 \seealso{\code{\link{readIntensities}},
          \code{\link{processArrayData}}
          }
 }

 \author{Andreas Buness <a.buness@dkfz.de>}

 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% End of File
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
