%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% This file was automatically generated by Rdoc on 2006-04-12  8:38:18.
% Do NOT modify this file, instead modify the source (/buness/R/TestingOfarrayMagic/Rpacks/arrayMagic/R/visualiseHybridisations.R).
% Rdoc is copyright (C) 2001, Henrik Bengtsson, henrikb@braju.com.
% Download the Rdoc compiler at http://www.braju.com/R/
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%


 \name{visualiseHybridisations}

 \title{Microarray intensity data are visualised with respect
        to their spatial layout}

 \alias{visualiseHybridisations}

 \usage{visualiseHybridisations(arrayDataObject, exprSetRGObject, type="raw", hybridisations, slideNameColumn, numberOfSpots, nrOfBlocksPerRow=4, mappingColumns= list(Block="Block", Column="Column", Row="Row"), transFunc=log, savePath=".", plotOutput="pdf")}

 \description{Two colour microarray intensity data are visualised
              with respect to their spatial layout.
              Missing values (\code{NA}) as well as
              minus infinity (\code{-Inf})
              etc. are not visualised.
              They appear to be white, i.e. in the colour of the background.
              \code{-Inf} might be a result of the (log-)transformation
              of the data; cf. the argument \code{transFunc}.}

 \value{None, but a plot is generated.}

 \arguments{

  \item{arrayDataObject}{object of class \code{\link{arrayData}};
                         required; default: missing}
  \item{exprSetRGObject}{object of class \code{\link{exprSetRG}};
                         only required,
                         if \code{type = "normalised"} or
                         \code{type = "normalisedLogRatio"};
                         default: missing}
  \item{type}{vector of character strings; required; default "raw";
              valid strings are "raw", "rawLogRatio",
              "rawLogRatioBackgroundSubtracted", "normalised"
              or "normalisedLogRatio"}
  \item{hybridisations}{vector of integers; optional; default: missing,
                        i.e. all microarray slides}
  \item{slideNameColumn}{character string specifying the column
                         \code{getHybAttr(arrayDataObject)}
                         which contains the names
                         of the hybridisations; optional;
                         default: missing}
  \item{numberOfSpots}{integer; optional; default: missing;
                       but required if the values in \code{arrayDataObject}
                       do not cover the whole microarray slide, i.e.
                       do not contain all spots;
                       cf. \code{\link{spatialLayout}}.}
  \item{nrOfBlocksPerRow}{integer; cf. \code{\link{spatialLayout}};
                          required; default: 4}
  \item{mappingColumns}{list; required; the list elements named
                        "Block", "Column", "Row" define the
                        corresponding column names given in
                        \code{getSpotAttr(arrayDataObject)};
                        the "Block" element is optional, if
                        missing a single block is assumed;
                        default: \code{list(Block="Block",
                        Column="Column", Row="Row")};
                        cf. \code{\link{spatialLayout}}
                       }
  \item{transFunc}{transformation function
                   applied to the raw data,
                   in particular in combination with \code{type = "raw"}
                   you may for example use \code{log} or \code{rank};
                   (note: \code{type = "rawLogRatio"} and
                          \code{type = "rawLogRatioBackgroundSubtracted"}
                          require log );
                   required; default: log}
  \item{savePath}{character string; required; default: "."}
  \item{plotOutput}{character string; required; either "screen", "pdf"
                    or "win.metafile"; default: "pdf"}

 }
 
 \examples{
	}

 \seealso{
           \code{\link{spatialLayout}},
           \code{\link{exprSetRG-class}},
           \code{\link{arrayData-class}}
         }

 \keyword{hplot}

 \author{Andreas Buness <a.buness@dkfz.de>}
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% End of File
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
