%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% This file was automatically generated by Rdoc on 2006-04-12  8:38:18.
% Do NOT modify this file, instead modify the source (/buness/R/TestingOfarrayMagic/Rpacks/arrayMagic/R/visualiseQualityParameters.R).
% Rdoc is copyright (C) 2001, Henrik Bengtsson, henrikb@braju.com.
% Download the Rdoc compiler at http://www.braju.com/R/
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%


 \name{visualiseQualityParameters}

 \title{Graphical representation of quality parameters}

 \alias{visualiseQualityParameters}

 \description{Graphical representation of quality parameters}

 \value{A graphical representation of quality parameters.
        \code{\link{qualityDiagnostics}} does call
        \code{visualiseQualityParameters} }

 \usage{visualiseQualityParameters(qualityParameters, savePath, fileName="visualiseQualityParametersOutput.pdf", width=25, height=25, plotOutput="pdf")}

 \arguments{
  \item{qualityParameters}{\code{data.frame}; required; 
                           contained in the return value of
                           \code{\link{qualityParameters}} }
  \item{fileName}{character string; required;
                  default: "visualiseQualityParametersOutput.pdf" }
  \item{savePath}{character string; optional; default: missing}
  \item{width}{graphics window width}
  \item{height}{graphics window height}
  \item{plotOutput}{character string; required; either "screen", "pdf"
                    or "win.metafile"; default: "pdf"}
 }
 
 \details{For details on the specific parameters
          see \code{\link{qualityParameters}}.
          Lines are only used for easier
          detection of outliers. The ordering
          of the hybridisation is somehow arbitrary,
          and only reflects the (initial) ordering. 
          The "correlation" line for example
          graphically represents the correlation
          coefficients calculated between the
          green and red channel for each hybridisation.
          The horizontal blue line is drawn
          at the height of one. The function
          \code{\link{qualityDiagnostics}} calls
          \code{visualiseQualityParameters}.
         }

 \seealso{\code{\link{qualityParameters}};
          \code{\link{qualityDiagnostics}} }

 \examples{
	}

 \keyword{hplot} ## required

 \author{Andreas Buness <a.buness@dkfz.de>}
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% End of File
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
